config

+++ src/main/java/com/cyanogen/cognition/block_entities/bibliophage/bookshelves/AbstractInfectedBookshelfEntity.java
@@ -27,7 +27,7 @@ import java.util.List;
 public abstract class AbstractInfectedBookshelfEntity extends AbstractInfectiveEntity {
 
     public AbstractInfectedBookshelfEntity(BlockEntityType<?> type, BlockPos pos, BlockState state,
-                                           int spawnDelayMin, int spawnDelayMax, int orbValue, int spawns, float infectivity) {
+                                           int spawnDelayMin, int spawnDelayMax, int orbValue, int spawns, double infectivity) {
 
         super(type, pos, state);
 
@@ -44,7 +44,7 @@ public abstract class AbstractInfectedBookshelfEntity extends AbstractInfectiveE
     final int orbValue; //the value of orbs to spawn
     final int spawns; //the number of times a bookshelf can spawn an orb before decaying
     int decayValue = 0; //the number of times a bookshelf has spawned an orb
-    final float infectivity; //the chance for a bookshelf to infect another adjacent bookshelf every second
+    final double infectivity; //the chance for a bookshelf to infect another adjacent bookshelf every second
     boolean redstoneEnabled = false; //whether the bookshelf is sensitive to redstone. Disabled bookshelves will not infect adjacents, produce XP, or decay
 
     public static final Component FROM_BOOKSHELF = Component.literal("SpawnedFromBookshelf");
